+++ sdk/cosmos/azure_data_cosmos_driver/src/models/mod.rs
@@ -29,11 +29,11 @@ mod user_agent;
 pub(crate) mod vector_session_token;
 pub(crate) use cosmos_headers::request_header_names;
 #[allow(dead_code)]
-pub(crate) mod effective_partition_key;
+pub mod effective_partition_key;
 #[allow(dead_code)]
 mod murmur_hash;
 #[allow(dead_code)]
-pub(crate) mod partition_key_range;
+pub mod partition_key_range;
 #[allow(dead_code)]
 pub(crate) mod range;